Add temporary logging for T210739
Attempts to reproduce the bug locally and on mwdebug1001 have failed, so
let's add some logging in production to try to get more information.

Note the LOCK IN SHARE MODE is there to try to bypass MySQL transaction
isolation, as that's one theory as to why things are screwing up.

Bug: T210739
Change-Id: I3d9b4ae4a274a0dee7d4c3361e548b90d4f21bba